Revert "Refactor backend"
