--- src/main/java/com/greenfox/szilvi/githubchecker/commitcheck/model/GfCommits.java
@@ -1,18 +1,0 @@
-package com.greenfox.szilvi.githubchecker.commitcheck.model;
-
-import lombok.Getter;
-
-@Getter
-public class GfCommits {
-
-    private String sha;
-    private Commit commit;
-
-    @Override
-    public String toString() {
-        return "GfCommits{" +
-                "sha='" + sha + '\'' +
-                ", commit=" + commit +
-                '}';
-    }
-}
