--- src/main/java/com/greenfox/szilvi/githubchecker/general/Settings.java
@@ -1,11 +1,0 @@
-package com.greenfox.szilvi.githubchecker.general;
-
-public class Settings {
-
-    public static String GITHUB_TOKEN = "accessToken";
-    public static boolean IS_LOCALHOST = true;
-    public static String GITHUB_ORG = "green-fox-academy";
-    public static String TODO_APP = "todo-app";
-    public static String WANDERER = "wanderer-";
-
-}
