--- src/main/java/com/greenfox/szilvi/githubchecker/greenfoxteam/model/GreenfoxTeamResponse.java
@@ -1,16 +1,0 @@
-package com.greenfox.szilvi.githubchecker.greenfoxteam.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-@Setter
-public class GreenfoxTeamResponse {
-
-    private int id;
-    private String name;
-}
