+++ src/main/java/com/greenfox/szilvi/githubchecker/models/TeamResponse.java
@@ -1,0 +1,16 @@
+package com.greenfox.szilvi.githubchecker.models;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Setter
+public class TeamResponse {
+
+    private int id;
+    private String name;
+}
