+++ src/main/java/com/greenfox/szilvi/githubchecker/services/Settings.java
@@ -1,0 +1,11 @@
+package com.greenfox.szilvi.githubchecker.services;
+
+public class Settings {
+
+    public static String GITHUB_TOKEN = "accessToken";
+    public static boolean IS_LOCALHOST = false;
+    public static String GITHUB_ORG = "green-fox-academy";
+    public static String TODO_APP = "todo-app";
+    public static String WANDERER = "wanderer-";
+
+}
